refactor: Row.java as inner class of Matrix for adjacency purposes, intermediate change
