+++ jogo-oito/src/main/java/factories/GameFactory.java
@@ -2,7 +2,6 @@ package factories;
 
 import interfaces.Edge;
 import interfaces.Vertex;
-import model.Row;
 
 public interface GameFactory {
 
@@ -12,7 +11,5 @@ public interface GameFactory {
     Edge createAdjacentDown(Vertex cell);
     Edge createAdjacentLeft(Vertex cell);
     Edge createAdjacentRight(Vertex cell);
-
-    Row createRow();
 
 }
