+++ jogo-oito/src/main/java/model/Matrix.java
@@ -14,35 +14,19 @@ public final class Matrix {
     private final Row thirdRow;
     public static List<Vertex> cells;
     //temp var
-    private final GameFactory gameFactory = new GameFactoryImpl();
 
     public Matrix() {
         Matrix.cells = new ArrayList<>();
         Cell.content = 1;
-        this.firstRow = gameFactory.createRow();
-        this.secondRow = gameFactory.createRow();
-        this.thirdRow = gameFactory.createRow();
+        this.firstRow = new Row();
+        this.secondRow = new Row();
+        this.thirdRow = new Row();
         this.defineAdjacent();
     }
 
     private void defineAdjacent() {
-        this.firstRow.initial.addAdjacents(gameFactory.createAdjacentUp(this.secondRow.initial));
-        this.secondRow.initial.addAdjacents(gameFactory.createAdjacentDown(this.firstRow.initial));
-
-        this.firstRow.center.addAdjacents(gameFactory.createAdjacentUp(this.secondRow.center));
-        this.secondRow.center.addAdjacents(gameFactory.createAdjacentDown(this.firstRow.center));
-
-        this.firstRow.last.addAdjacents(gameFactory.createAdjacentUp(this.secondRow.last));
-        this.secondRow.last.addAdjacents(gameFactory.createAdjacentDown(this.firstRow.last));
-
-        this.secondRow.initial.addAdjacents(gameFactory.createAdjacentUp(this.thirdRow.initial));
-        this.thirdRow.initial.addAdjacents(gameFactory.createAdjacentDown(this.secondRow.initial));
-
-        this.secondRow.center.addAdjacents(gameFactory.createAdjacentUp(this.thirdRow.center));
-        this.thirdRow.center.addAdjacents(gameFactory.createAdjacentDown(this.secondRow.center));
-
-        this.secondRow.last.addAdjacents(gameFactory.createAdjacentUp(this.thirdRow.last));
-        this.thirdRow.last.addAdjacents(gameFactory.createAdjacentDown(this.secondRow.last));
+        this.firstRow.defineVerticalAdjacency(this.secondRow);
+        this.secondRow.defineVerticalAdjacency(this.thirdRow);
 
         this.changePositionToValidateTemplate();
     }
@@ -53,5 +37,45 @@ public final class Matrix {
 
     public List<Vertex> getCells() {
         return Matrix.cells;
+    }
+
+    public static final class Row {
+
+        public final Vertex initial;
+        public final Vertex center;
+        public final Vertex last;
+        private final GameFactory gameFactory = new GameFactoryImpl();
+
+        public Row() {
+            this.initial = gameFactory.createCell();
+            this.center = gameFactory.createCell();
+            this.last = gameFactory.createCell();
+            this.defineHorizontalAdjacency();
+            this.loadCells();
+        }
+
+        public void loadCells() {
+            Matrix.cells.add(this.initial);
+            Matrix.cells.add(this.center);
+            Matrix.cells.add(this.last);
+        }
+
+        public void defineVerticalAdjacency(Row row) {
+            this.initial.addAdjacents(gameFactory.createAdjacentUp(row.initial));
+            this.center.addAdjacents(gameFactory.createAdjacentUp(row.center));
+            this.last.addAdjacents(gameFactory.createAdjacentUp(row.last));
+
+            row.initial.addAdjacents(gameFactory.createAdjacentDown(this.initial));
+            row.center.addAdjacents(gameFactory.createAdjacentDown(this.center));
+            row.last.addAdjacents(gameFactory.createAdjacentDown(this.last));
+        }
+
+        public void defineHorizontalAdjacency() {
+            this.initial.addAdjacents(gameFactory.createAdjacentLeft(this.center));
+            this.center.addAdjacents(gameFactory.createAdjacentRight(this.initial));
+            this.center.addAdjacents(gameFactory.createAdjacentLeft(this.last));
+            this.last.addAdjacents(gameFactory.createAdjacentRight(this.center));
+        }
+
     }
 }
